working on the menu
