+++ src/js/elements/point_tray/styled_elements.js
@@ -1,5 +1,6 @@
 import {styled} from "@mui/system";
 import TreeItem, {treeItemClasses} from "@mui/lab/TreeItem";
+import {Switch} from "@mui/material";
 
 export const CategoryTreeItem = styled(TreeItem)(() => ({
     [`& .${treeItemClasses.content}`]: {
@@ -40,8 +41,51 @@ export const ThemeTreeItem = styled(TreeItem)(() => ({
         [`& .${treeItemClasses.label}`]: {
             fontWeight: '500',
             fontSize: '0.9em',
-            padding: '11px',
+            padding: '8px',
             lineHeight: '1.2'
         }
+    },
+    [`& .${treeItemClasses.selected}`]: {
     }
+}));
+
+export const AntSwitch = styled(Switch)(({theme}) => ({
+    width: 32,
+    height: 20,
+    padding: 0,
+    margin: '0 auto',
+    display: 'flex',
+    '&:active': {
+        '& .MuiSwitch-thumb': {
+            width: 15,
+        },
+        '& .MuiSwitch-switchBase.Mui-checked': {
+            transform: 'translateX(9px)',
+        },
+    },
+    '& .MuiSwitch-switchBase': {
+        padding: 2,
+        '&.Mui-checked': {
+            transform: 'translateX(12px)',
+            color: '#fff',
+            '& + .MuiSwitch-track': {
+                opacity: 1,
+                backgroundColor: theme.palette.mode === 'dark' ? '#177ddc' : '#1890ff',
+            },
+        },
+    },
+    '& .MuiSwitch-thumb': {
+        boxShadow: '0 2px 4px 0 rgb(0 35 11 / 20%)',
+        width: 12,
+        height: 12,
+        borderRadius: 6,
+        marginTop: 2,
+        marginLeft: 2
+    },
+    '& .MuiSwitch-track': {
+        borderRadius: 34,
+        opacity: 1,
+        backgroundColor: 'rgba(0,0,0,.18823529411764706)',
+        boxSizing: 'border-box',
+    },
 }));
